fixed (#14274)

+++ samples/client/petstore/csharp-netcore/OpenAPIClient-generichost-net6.0-nrt/src/Org.OpenAPITools/Model/ApiResponse.cs
@@ -179,7 +179,7 @@ namespace Org.OpenAPITools.Model
         {
             writer.WriteStartObject();
 
-            writer.WriteNumber("code", (int)apiResponse.Code);
+            writer.WriteNumber("code", apiResponse.Code);
             writer.WriteString("message", apiResponse.Message);
             writer.WriteString("type", apiResponse.Type);
