+++ samples/client/petstore/csharp-netcore/OpenAPIClient-generichost-net6.0/src/Org.OpenAPITools/Model/Banana.cs
@@ -145,7 +145,7 @@ namespace Org.OpenAPITools.Model
         {
             writer.WriteStartObject();
 
-            writer.WriteNumber("lengthCm", (int)banana.LengthCm);
+            writer.WriteNumber("lengthCm", banana.LengthCm);
 
             writer.WriteEndObject();
         }
